engines/symbapt/pyproject.toml
[project]
name = "symbapt"
version = "0.1.0"
description = "Neurosymbolic APT hunter — differentiable MITRE ATT&CK rules + Kafka streaming"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
    "flask>=3.0",
]

[project.optional-dependencies]
kafka = ["kafka-python>=2.0"]
dev   = ["pytest", "pytest-cov"]

[project.scripts]
symbapt = "engines.symbapt.server:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"
